More function prototype fixes
svn path=/trunk/; revision=27972

+++ reactos/dll/cpl/ncpa/ncpa.h
@@ -13,7 +13,7 @@ typedef struct
 
 extern HINSTANCE hApplet;
 
-extern DWORD DbgPrint(PCCH Fmt, ...);
+extern ULONG DbgPrint(PCCH Fmt, ...);
 
 #endif // __NCPA_H
